requirements.txt
streamlit>=1.20
sentence-transformers>=2.2.2
faiss-cpu              # or faiss-gpu if you have GPU
transformers>=4.30.0
torch>=1.12
openai>=0.27.0         # optional, only if you use OpenAI API
tqdm
python-multipart
pdfminer.six>=20201018 # for PDF extraction
